[remove] CommonButton 제거하고 button에 통합 #84
