+++ src/components/navigation/NavSidebar.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 import { TbFolders, TbFolderShare, TbSmartHome } from 'react-icons/tb';
 
-import { Button } from '@/components/buttons/CommonButton';
+import { Button } from '@/components/ui/button';
 import { ScrollArea } from '@/components/ui/scroll-area';
 import { Separator } from '@/components/ui/separator';
